Add a blob existence cache to RemoteCache
Bazel currently performs a full findMissingBlob call for every input of every
action that is not an immediate cache hit. This causes a significant amount
of network traffic as well as server load (unless the server caches blob
existence as well).

In particular, for a build of TensorFlow w/ remote execution over a slow-ish
network, we see multi-minute delays on actions due to these findMissingBlob
calls, and the client is not able to saturate a 50-machine cluster.

This change carefully de-duplicates findMissingBlob calls as well as file
uploads to the remote execution cluster. It is based on a previous simpler
change that did not de-duplicate *concurrent* calls. However, we have found
that concurrent calls are quite common - i.e., an action completes that
unblocks a large number of actions that have significant overlaps in their
input sets (e.g., protoc link completion unblocks all protoc compile
actions), and we were still seeing multi-minute delays on action execution.

With this change, a TensorFlow build w/ remote execution is down to ~20
minutes on a 50-machine cluster, and is able to fully saturate the cluster
(apart from an issue with connection loss, see PR #11957).

Change-Id: Ic39347a7a7a8dc7cfd463d78f0a80e0d26a970bc